backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
google-generativeai==0.8.3
